+++ test.sh
@@ -1,4 +1,4 @@
-cd dio
+cd dio || exit
 dart test --coverage=coverage .
 pub run coverage:format_coverage --packages=.packages -i coverage -o coverage/lcov.info --lcov
 genhtml -o coverage coverage/lcov.info