+++ plugins/http2_adapter/CHANGELOG.md
@@ -1,3 +1,6 @@
+## NEXT
+
+* Add `validateCertificate` for `ClientSetting`.
 
 ## [2.0.0]